add files to cocoapods project
