--- KarhooUISDK/Screens/CheckoutScreen/Components/File.swift
@@ -1,19 +1,0 @@
-//
-//  KarhooAddFlightNumberViewMVP.swift
-//
-//
-//  Created by Bartlomiej Sopala on 05/01/2023.
-//
-
-import Foundation
-import KarhooSDK
-
-protocol AddFlightNumberViewDelegate: AnyObject {
-    func willUpdateFlightNumber()
-    func didUpdateFlightNumber(_ flightNumber: String)
-}
-
-protocol AddFlightNumberPresenter {
-    func set(flightNumber: String)
-}
-
