+++ Podfile
@@ -1,5 +1,5 @@
 # Uncomment the next line to define a global platform for your project
-platform :ios, '13.0'
+platform :ios, '14.0'
 
 # Standard cocoapods specs source
 source 'https://cdn.cocoapods.org/'
